rename: presentations/shared を _shared に変更
