+++ presentations/jsconfjp-2021/.eleventy.js
@@ -1,4 +1,4 @@
-import { configureEleventy } from "../shared/eleventy-base.js";
+import { configureEleventy } from "../_shared/eleventy-base.js";
 
 export default function (eleventyConfig) {
   return configureEleventy(eleventyConfig, {
